+++ demos/demo33/application.c
@@ -1,0 +1,29 @@
+/**
+ * File:   application.c
+ * Author: AWTK Develop Team
+ * Brief:  application
+ *
+ * Copyright (c) 2020 - 2020  Guangzhou ZHIYUAN Electronics Co.,Ltd.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * License file for more details.
+ *
+ */
+
+/**
+ * History:
+ * ================================================================
+ * 2020-12-06 Li XianJing <xianjimli@hotmail.com> created
+ *
+ */
+
+#include "awtk.h"
+#include "mvvm/mvvm.h"
+
+ret_t application_init(void) {
+  return navigator_to("demo33");
+}
+
+#include "../awtk_main.c"